Finished moving to proxy/policy, moved to proxy/utils/[tf_ops, tracking_dict]
