--- python/ray/rllib/agents/sac/sac/rllib_proxy/_constants.py
@@ -1,3 +1,0 @@
-# Default policy id for single agent environments
-DEFAULT_POLICY_ID = "default_policy"
-
